+++ packages/core/src/node/initRsbuild.ts
@@ -1,6 +1,6 @@
 import path from 'path';
 import type { UserConfig } from '@rspress/shared';
-import fs from '@modern-js/utils/fs-extra';
+import fs from '@rspress/shared/fs-extra';
 import {
   MDX_REGEXP,
   RSPRESS_TEMP_DIR,
@@ -40,7 +40,7 @@ async function createInternalBuildConfig(
   pluginDriver: PluginDriver,
 ): Promise<RsbuildConfig> {
   const cwd = process.cwd();
-  const { default: fs } = await import('@modern-js/utils/fs-extra');
+  const { default: fs } = await import('@rspress/shared/fs-extra');
   const CUSTOM_THEME_DIR =
     config?.themeDir ?? path.join(process.cwd(), 'theme');
   const outDir = config?.outDir ?? OUTPUT_DIR;
@@ -109,6 +109,7 @@ async function createInternalBuildConfig(
       template: path.join(PACKAGE_ROOT, 'index.html'),
     },
     output: {
+      targets: isSSR ? ['node'] : ['web'],
       distPath: {
         // `root` must be a relative path in Rsbuild
         root: path.isAbsolute(outDir) ? path.relative(cwd, outDir) : outDir,
@@ -244,7 +245,6 @@ export async function initRsbuild(
   );
 
   const rsbuild = await createRsbuild({
-    target: isSSR ? 'node' : 'web',
     rsbuildConfig: mergeRsbuildConfig(
       internalRsbuildConfig,
       ...(config?.plugins?.map(plugin => plugin.builderConfig ?? {}) || []),