+++ packages/core/src/node/utils/flattenMdxContent.ts
@@ -1,7 +1,7 @@
 import path from 'path';
 import enhancedResolve from 'enhanced-resolve';
 import type { Resolver } from 'enhanced-resolve';
-import fs from '@modern-js/utils/fs-extra';
+import fs from '@rspress/shared/fs-extra';
 import { createProcessor } from '@mdx-js/mdx';
 import type { Root } from 'hast';
 import { MDX_REGEXP } from '@rspress/shared';