+++ packages/shared/chalk.d.ts
@@ -1,0 +1,2 @@
+export * from './dist/chalk';
+export { default } from './dist/chalk';