+++ packages/plugin-shiki/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rspress/plugin-shiki
 
+## 1.8.3
+
+### Patch Changes
+
+- Updated dependencies [d4f89b3]
+- Updated dependencies [bfe47e8]
+- Updated dependencies [38a2c6e]
+  - @rspress/shared@1.8.3
+
 ## 1.8.2
 
 ### Patch Changes