+++ packages/plugin-typedoc/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @rspress/plugin-typedoc
 
+## 1.8.3
+
+### Patch Changes
+
+- 38a2c6e: perf: optimize dependencies size
+- Updated dependencies [d4f89b3]
+- Updated dependencies [bfe47e8]
+- Updated dependencies [38a2c6e]
+  - @rspress/shared@1.8.3
+
 ## 1.8.2
 
 ## 1.8.1